sagemaker/requirements.txt
transformers>=4.43
datasets>=2.18
accelerate>=0.31
evaluate>=0.4.1
torch>=2.3
pillow>=10.3
timm>=0.9.16
sacrebleu>=2.4.2
requests>=2.32
